Adds an IBAN type with relaxed validation constraints.
This might be useful in some cases, e.g. when handling input.  LenientIBAN may
not be the best of names, but the term "Lenient" matches the relaxed validation
toggle in java.util.Calendar, so it's not completely foreign. I wanted to keep
the simple, straightforward IBAN moniker for the strictly validated instances.

Relaxed validation instances (which may in fact be invalid) cannot be assigned
to variables of the stricter IBAN type, but the other way around is no problem.