use code generator for reference data
This splits the `CountryCodes` class into `CountryCodes` (hand-written) and
`CountryCodesData` (generated). The input for the code generator is a YAML file
in `src/main/resources`. It is added to the distribution JAR file, but it is
not used at runtime.

This makes it a little easier to maintain, not having to match the item
ordering in three different arrays. The reference data in
`CountryCodesParameterizedTest` is maintained separately, to enforce a sort of
double-check of my own work.